content-copilot/core/scripts/toc-test-2.py
#!/usr/bin/env python3
"""
TOC Generation Script

Este script gera propostas de TOC (Table of Contents) usando prompts de IA.

Uso:
    python core/scripts/toc-test-2.py --input prompt_bundle.json --output toc_proposals.json
"""

import argparse
import json
import yaml
from typing import Dict, List
from dotenv import load_dotenv
import os


def load_prompt_bundle(bundle_path: str) -> Dict:
    """
    Carrega um bundle de prompt (ex.: prompt.yaml, templates, schemas).
    
    Args:
        bundle_path (str): Caminho para o arquivo de bundle.
    
    Returns:
        Dict: Dicionário com o bundle carregado.
    """
    with open(bundle_path, 'r', encoding='utf-8') as f:
        return json.load(f)


def generate_toc_proposals(bundle: Dict) -> Dict:
    """
    Gera propostas de TOC usando um bundle de prompt.
    
    Args:
        bundle (Dict): Bundle de prompt carregado.
    
    Returns:
        Dict: Dicionário com as propostas de TOC geradas.
    """
    # Simulação: Gerar TOCs estáticas para teste
    # Em uma implementação real, esta função chamaria a API de IA
    proposals = [
        {
            "title": "Guia Completo sobre Inteligência Artificial",
            "toc": [
                {
                    "level": "h1",
                    "text": "Introdução à Inteligência Artificial"
                },
                {
                    "level": "h2",
                    "text": "Aplicações da IA",
                    "children": [
                        {
                            "level": "h3",
                            "text": "Saúde"
                        },
                        {
                            "level": "h3",
                            "text": "Finanças"
                        }
                    ]
                }
            ]
        },
        {
            "title": "Como a IA Está Transformando os Negócios",
            "toc": [
                {
                    "level": "h1",
                    "text": "Impacto da IA nos Negócios"
                },
                {
                    "level": "h2",
                    "text": "Automação de Processos"
                },
                {
                    "level": "h2",
                    "text": "Tomada de Decisão Baseada em Dados"
                }
            ]
        },
        {
            "title": "O Futuro da Inteligência Artificial",
            "toc": [
                {
                    "level": "h1",
                    "text": "Tendências em IA"
                },
                {
                    "level": "h2",
                    "text": "IA Generativa"
                },
                {
                    "level": "h2",
                    "text": "Ética e Responsabilidade"
                }
            ]
        }
    ]
    
    return {"proposals": proposals}


def save_output(output: Dict, output_path: str) -> None:
    """
    Salva as propostas de TOC em um arquivo JSON.
    
    Args:
        output (Dict): Dicionário com as propostas de TOC.
        output_path (str): Caminho para o arquivo de saída.
    """
    with open(output_path, 'w', encoding='utf-8') as f:
        json.dump(output, f, ensure_ascii=False, indent=2)


def main():
    """
    Função principal para gerar TOCs.
    """
    parser = argparse.ArgumentParser(description='Gera propostas de TOC usando prompts de IA.')
    parser.add_argument('--input', type=str, required=True, help='Caminho para o arquivo de bundle de prompt.')
    parser.add_argument('--output', type=str, required=True, help='Caminho para o arquivo JSON de saída.')
    
    args = parser.parse_args()
    
    # Carregar bundle de prompt
    bundle = load_prompt_bundle(args.input)
    
    # Gerar propostas de TOC
    toc_proposals = generate_toc_proposals(bundle)
    
    # Salvar saída
    save_output(toc_proposals, args.output)
    
    print(f"Propostas de TOC geradas e salvas em: {args.output}")


if __name__ == '__main__':
    main()