content-copilot/requirements.txt
# Dependências Python para o CITO MVP
#
# Instale as dependências com:
# pip install -r requirements.txt

# Core Dependencies
beautifulsoup4==4.12.2
jinja2==3.1.2
pymongo==4.6.1
pyyaml==6.0.1
requests==2.31.0
jsonschema==4.19.0
python-dotenv==1.0.0

# Logging
loguru==0.7.0

# Testing (Optional)
pytest==7.4.0
pytest-mock==3.11.1